Major timer-based refactor
All sub-processes have been removed and replaced with a timer in the
main process. This timer fires every 20 ms or so and retrieves audio
data. This removes a lot of complexity coming from synchronization
issues.

Additionnally a level-meter widget is introduced, and the plot widgets
have been optimized so that only the needed part is redrawn (namely axes
are not redrawn).