Windows build: Create a build script, switch to PyPI PyQt5 wheels
The build script is written in Powershell, with a thin cmd wrapper too.
It installs and builds everything in a virtualenv, up the final NSIS
installer.

This allows to run the same build locally as Appveyor will do.

The PyPI PyQt5 wheels install the Qt dlls in a place that PyInstaller does
not know, so a workaround is added to the spec file.