Replace NSIS installer by MSI generated with WiX
NSIS installer has limitations. The biggest one is that the uninstaller is quite dumb and just deletes the install folder instead of tracking which files are installed. This can be dramatic if the user happens to install in a folder that already contains other files, leading to loss of data when uninstalling.

Here we replace the NSIS Installer by an MSI generated by the WiX toolset. It is more robust and retains similar features (choice of install location, shortcut, license page).