Use python-rtmixer for low-latency audio recording
Friture has issues keeping up with audio input, which tends to produce overflows, that appear as glitches in the audio signal.

This has largely been unnoticed for years because Windows MME is very forgiveable in that regard. But Windows Directsound, for example, is much more sensitive. Similarly, Alsa or PulseAudio in Linux also show a lot of glitches (even if PortAudio might not report any overflow).

Here this is fixed vy taking advantage of python-rtmixer, a library that sits on top of python-sounddevice (the PortAudio wrapper). Python-rtmixer provides an audio callback that is implemented in C (and compiled with the help of CFFI) and doesn’t invoke the Python interpreter, therefore avoiding waiting for things like garbage collection and the GIL.

Reference: https://python-rtmixer.readthedocs.io/en/0.1.0/index.html

The result is glitch-free audio signal even with Windows DirectSound.

We believe this solves #90