Log scale minor ticks: show both odd and even divisions
Friture only has the 'even' division lines showing in “Logarithmic” Frequency scale for the FFT Spectrum. For example, the frequency division lines for 3000, 5000, 7000 and 9000 Hz are not shown on the Frequency scale when it’s set to “Logarithmic” mode. However, all of the graph’s division lines are present when set to “Linear” mode.

It can be a bit misleading.

Here we add the odd divisions too. It makes the graph slightly more cluttered, but favors visual accuracy and least surprise.

Fixes #47