Remove incorrect decimation in FFT spectrum
Addresse #132. When the FFT spectrum is cropped by lowering the max frequency, the signal was decimated N times by taking one sample every N samples. This is not correct, produces many artifacts. To keep decimation, we would need to use a proper low-pass filter (as implemented for the octave spectum for example), but it does not really seem worth it. We will just apply the FFT on top of the undecimated signal.