Macos builds: install latest portaudio
For Macos Big Sur, the stable portaudio (19.6.0) makes Friture freeze on startup.
So we install from latest master instead, and we tell pyinstaller to bundle that version instead of the version provided by python-sounddevice.
See #154

Also remove the vcruntime140.dll workaround that is not needed with latest pyinstaller, and was getting confusing to apply next to the Macos setting.