Restore state of previously closed docks in new docks.
This enables the following:
- add a dock
- select a different widget
- change some settings
- close that dock
- add a new dock
- new dock has the widget and settings of the one previously closed

Previous widget selections are kept on a stack, so this allows
"new dock" to undo multiple previous "close" actions. Caveat being that
if multiple instances of the same widget were closed, only the most
recent one's settings will be kept.

This isn't persisted at close, yet.