chore: move winbuild scripts to github folder
Apparently the winbuild script at the root of the repo was confusing, it looked like it was meant to be used to run Friture locally from source. However, it's actually meant to be used from the Github action. So we move it the .github folder to make it clearer. This actually harmonizes the file structure with linux and macos.