feat: migrate spectrogram to QtQuick/QML
This is the last widget to be migrated. The benefits have to do with the easier layout code mostly.

For now, this is implemented with QQuickPaintedItem to expedite the migration, so there is not much performance benefit (10% less CPU nevertheless), but this could be revisited if one day we move to Qt6 to use a texture node in the scene graph.

Also, there could be a round of extra refactoring in the future to remove some of the classes, the current state has a bit of redundancy.