feat(QML): all widgets are now QML components
All audio widgets are now QML components. This required migrating the delay estimator, the generator, and refactoring the Pitch tracker.

The benefit is that now the creation of the visual part of the audio widget is centralized in dock.py, instead of being copied over in every widget. Also, it promotes a clearer separation between viewmodel and view.

This will allow to continue the conversion of Friture to be a full QML app.

We choose the Fusion style for Qt Quick Controls, used in the Generator view. The Fusion style is appropriate for Desktop apps, and works fine with Dark or White themes as set in the system.